Fix wrong quoting characters for SOSL FIND queries
